+++ packages/core/standard/src/ed2/rev1/unstable/IEC61850-90-30/definition/substation/types.substation.d.ts
@@ -11,6 +11,8 @@ export namespace Substation {
 	export type LNode = {
 		tag: 'LNode'
 		attributes: {
+			uuid: string
+			templateUuid?: string
 			originUuid?: string
 		}
 	}