+++ packages/plugins/type-designer/src/headless/stores/index.ts
@@ -2,6 +2,7 @@ export * from './dnd.svelte'
 export * from './plugin-local.svelte'
 export * from './sidebar.svelte'
 export * from './type-elements/store.svelte'
-
+export * from './imports/store.svelte'
 // types
 export type * from './type-elements/types'
+export type * from './imports/types'