+++ packages/plugins/type-designer/src/ui/components/sidebar-wrapper.svelte
@@ -13,6 +13,10 @@ import { TYPE_FAMILY, READONLY_ATTRIBUTES } from '@/headless/constants'
 // STORES
 import { sidebarStore, pluginLocalStore } from '@/headless/stores'
 
+//====== CONSTANTS ======//
+
+const attributesNeedingOneLineLayout = ['virtual']
+
 //====== REACTIVE VARIABLES ======//
 
 const currentElementAttributesEntries = $derived.by(() => {
@@ -34,12 +38,11 @@ const isAnyAttributeAllowed = $derived.by(() => {
 
 //====== FUNCTIONS ======//
 
-function isAttributeReadonly(attributeKey: string) {
-	return READONLY_ATTRIBUTES.some((attributes) => attributes === attributeKey)
-}
-
-function onChangeHandler() {
-	sidebarStore.currentElementType &&
+function onChangeHandler(attributeKey: string) {
+	if (
+		sidebarStore.currentElementType &&
+		sidebarStore.isInputValidByAttributeKey?.[attributeKey]
+	)
 		pluginGlobalStore.updateElement({
 			element: sidebarStore.currentElementType.element,
 			attributes: sidebarStore.currentElementType.attributes
@@ -55,10 +58,19 @@ function setVirtualValue(value: boolean) {
 		sidebarStore.currentElementType.attributes.virtual = value
 			? 'true'
 			: 'false'
-	onChangeHandler()
+	onChangeHandler('virtual')
 }
 
-const attributesNeedingOneLineLayout = ['virtual']
+function isCurrentInputDisabled(attributeKey: string) {
+	return (
+		isAttributeReadonly(attributeKey) ||
+		sidebarStore.isCurrentElementImported
+	)
+}
+
+function isAttributeReadonly(attributeKey: string) {
+	return READONLY_ATTRIBUTES.some((attributes) => attributes === attributeKey)
+}
 </script>
 
 <Sidebar.Root side="right" class="sidebar-root z-0">
@@ -75,7 +87,7 @@ const attributesNeedingOneLineLayout = ['virtual']
 								class="mt-2"
 								bind:value={sidebarStore.currentElementType.attributes[attributeKey]}
 							/>
-						{:else if sidebarStore.currentElementTypeKey && sidebarStore.currentElementTypeFamily}
+						{:else if sidebarStore.currentElementTypeKey && sidebarStore.currentElementTypeFamily && sidebarStore.isInputValidByAttributeKey}
 							{#if attributeKey === 'virtual'}
 								<Switch.Root
 									class="mt-1 self-end"
@@ -83,19 +95,23 @@ const attributesNeedingOneLineLayout = ['virtual']
 										getVirtualValue,
 										setVirtualValue
 									}
-									oninput={onChangeHandler}
+									oninput={() => onChangeHandler(attributeKey)}
+									disabled={sidebarStore.isCurrentElementImported}
 								/>
 							{:else}
 								<Input.Root 
-									disabled={isAttributeReadonly(attributeKey)}
-									class="mt-1"
+									disabled={isCurrentInputDisabled(attributeKey)}
+									class={`mt-1 ${!isCurrentInputDisabled(attributeKey) && !sidebarStore.isInputValidByAttributeKey[attributeKey] ? 'border-destructive focus-visible:ring-destructive' : ''}`}
 									required={attributeProperties.required}
 									type="text"
 									id={attributeKey}
 									placeholder={isAttributeReadonly(attributeKey) ? '' : attributeKey}
 									bind:value={sidebarStore.currentElementType.attributes[attributeKey]}
-									onchange={onChangeHandler}
+									onchange={() => onChangeHandler(attributeKey)}
 								/>
+								{#if !isCurrentInputDisabled(attributeKey) && !sidebarStore.isInputValidByAttributeKey[attributeKey]}
+									<p class="text-destructive text-xs mt-1">This value should be unique</p>
+								{/if}
 							{/if}
 						{/if}
 					</div>